requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
authlib==1.6.0
backoff==2.2.1
beautifulsoup4==4.13.4
browser-use==0.4.0
bubus==1.2.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
distro==1.9.0
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-genai==1.22.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
groq==0.29.0
grpcio==1.73.1
grpcio-status==1.73.1
gspread==6.2.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.10.0
-e file:///home/ray/code/2025/JobApp
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-anthropic==0.3.16
langchain-core==0.3.66
langchain-google-genai==2.1.5
langchain-openai==0.3.26
langchain-text-splitters==0.3.8
langgraph==0.5.0
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.1
langgraph-sdk==0.1.72
langsmith==0.4.3
markdownify==1.1.0
markupsafe==3.0.2
mem0ai==0.1.112
numpy==2.3.1
oauth2client==4.1.3
oauthlib==3.3.1
openai==1.92.2
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.0
patchright==1.52.5
platformdirs==4.3.8
playwright==1.53.0
pluggy==1.6.0
portalocker==2.10.1
posthog==6.0.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyee==13.0.0
pygments==2.19.2
pypandoc==1.15
pyparsing==3.2.3
pyperclip==1.9.0
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.14.3
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.41
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uuid7==0.1.0
websockets==15.0.1
xxhash==3.5.0
zstandard==0.23.0
